requirements.txt
streamlit
numpy
pandas
pillow
torch
torchvision
albumentations
opencv-python-headless
typing-extensions
dataclasses
ttach
effdet
